requirements.txt
opencv-python
ultralytics
pillow
numpy
imageio
